Updated main.py, datatier.py, modify_password.py, sign_in.py, and sign_up.py
