+++ datatier.py
@@ -1,0 +1,65 @@
+###################################################################
+#
+# web_service_get
+#
+# When calling servers on a network, calls can randomly fail. 
+# The better approach is to repeat at least N times (typically 
+# N=3), and then give up after N tries.
+#
+
+import logging
+import requests
+import time
+
+def web_service_get(url):
+    """
+    Submits a GET request to a web service at most 3 times, since
+    web services can fail to respond e.g. to heavy user or internet
+    traffic. If the web service responds with status code 200, 400
+    or 500, we consider this a valid response and return the response.
+    Otherwise we try again, at most 3 times. After 3 attempts the
+    function returns with the last response.
+
+    Parameters
+    ----------
+    url: url for calling the web service
+
+    Returns
+    -------
+    response received from web service
+    """
+
+    try:
+        retries = 0
+
+        while True:
+            response = requests.get(url)
+
+            if response.status_code in [200, 400, 480, 481, 482, 500]:
+                #
+                # we consider this a successful call and response
+                #
+                break;
+
+            #
+            # failed, try again?
+            #
+            retries = retries + 1
+            if retries < 3:
+                # try at most 3 times
+                time.sleep(retries)
+                continue
+
+            #
+            # if get here, we tried 3 times, we give up:
+            #
+            break
+
+        return response
+
+    except Exception as e:
+        print("**ERROR**")
+        logging.error("web_service_get() failed:")
+        logging.error("url: " + url)
+        logging.error(e)
+        return None
