+++ modify_password.py
@@ -1,0 +1,2 @@
+def modify_pwd():
+    return
